+++ src/app/constants/version.constants.ts
@@ -1,2 +1,2 @@
-export const version = "alpha 0.1.4";
-export const date = "2024-02-03";
+export const version = "alpha 0.1.5";
+export const date = "2024-02-07";